requirements.txt
# Pipeline deps (existing)
pandas>=2.0
numpy>=1.26
pyarrow>=14.0
requests>=2.31
lightgbm>=4.0
scikit-learn>=1.4
joblib>=1.3
torch>=2.1
python-dotenv>=1.0

# Web app
flask>=3.0
flask-sqlalchemy>=3.1
psycopg2-binary>=2.9
gunicorn>=21.2
apscheduler>=3.10
psutil>=5.9
